feat: create message presentation layer
